machine-learning-devops-engineer/predict-customer-churn/pyproject.toml
[tool.poetry]
name = "predict_customer_churn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "0.24.1"
shap = "0.40.0"
joblib = "1.0.1"
pandas = "1.2.4"
numpy = "1.20.1"
matplotlib = "3.3.4"
seaborn = "0.11.2"
pylint = "2.7.4"
autopep8 = "1.5.6"
pytest = "^7.1.2"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
